build config updated
