+++ src/index.tsx
@@ -1,4 +1,3 @@
 export { CadViewer } from "./CadViewer.tsx"
-export * from "./convert-circuit-json-to-3d-svg.tsx"
+export * from "./convert-circuit-json-to-3d-svg.ts"
 export * from "./hooks/index.ts"
-
