+++ tsup.config.ts
@@ -1,0 +1,14 @@
+import { defineConfig } from "tsup"
+
+export default defineConfig({
+  entry: ["src/index.tsx"],
+  format: ["cjs", "esm"],
+  dts: true,
+  sourcemap: true,
+  clean: true,
+  external: ["three"],
+  treeshake: true,
+  esbuildOptions(options) {
+    options.conditions = ["module"]
+  },
+})
